Cria is_ahead_of_print em SPS_Package e move build_xml p/ tests.utils
